+++ customize.sh
@@ -1,5 +1,5 @@
 SKIPUNZIP=1
-MODLOG=/data/high_perf_dac
+MODLOG=/storage/emulated/0/high_perf_mode.log
 MODEL=$(getprop ro.product.system.model 2>/dev/null)
 RELEASE=$(getprop ro.system.build.version.release 2>/dev/null)
 HARDWARE=$(getprop ro.hardware 2>/dev/null)
@@ -11,12 +11,12 @@ rm -rf $MODPATH/README.md 2>/dev/null
 }
 SET_PERMISSION() {
 set_perm_recursive $MODPATH 0 0 0755 0644
-set_perm $MODPATH/post-fs-data.sh 0 0 0777 0777
+set_perm $MODPATH/service.sh 0 0 0777 0777
 }
 MOD_EXTRACT() {
 ui_print "- Extracting module files"
 unzip -o "$ZIPFILE" module.prop -d $MODPATH >&2
-unzip -o "$ZIPFILE" post-fs-data.sh -d $MODPATH >&2
+unzip -o "$ZIPFILE" service.sh -d $MODPATH >&2
 }
 MOD_FUNCTION() {
 if [ "$HARDWARE" == "qcom" ]; then
@@ -47,10 +47,6 @@ ui_print ""
 MOD_EXTRACT
 MOD_FUNCTION
 }
-if [ ! "$SKIPUNZIP" = "0" ]; then
-    set -x
-    MOD_PRINT
-else
-    set +x
-    abort
-fi
+set -x
+MOD_PRINT
+